rename: station 도메인 파일 위치 수정
