+++ src/main/java/subway/station/Station.java
@@ -1,6 +1,10 @@
-package subway;
+package subway.station;
 
-import javax.persistence.*;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
 
 @Entity
 public class Station {
