+++ src/main/java/subway/station/StationController.java
@@ -1,10 +1,15 @@
-package subway;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+package subway.station;
 
 import java.net.URI;
 import java.util.List;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 public class StationController {
